Split relational storage DML
